refactor(setting): extract password change/set logic to ChangeOrSetPasswordListItem
Move password change/set dialog logic from config.tsx to a dedicated
ChangeOrSetPasswordListItem component in CustomElement.tsx for better
code organization and reusability.
